Precalculate buffer size in base64 functions
Aide uses a fixed size (16k bytes) for the return buffer in
encode_base64/decode_base64 functions. This results in a segfault if
aide processes a file with too large extended attribute value or ACL.

Fix this issue by precalculating the size of the return buffer depending on
the input in the encode_base64/decode_base64 functions.

This addresses CVE-2021-45417. Thanks to David Bouman for reporting this
vulnerability and reviewing this patch.